Add proof-of-concept for support for SMTP in-band TLS upgrade (STARTTLS).
This is implemented as a TransportSocket that wraps a RawBufferSocket
and a SslSocket and watches the protocol exchange until the client says
STARTTLS and then swaps the sockets out. The smtp inspection code is
currently inlined into the transport socket implementation. It isn't
clear to me if making this pluggable to share with other protocols is a
better or worse approach than doing one-offs for each one. It may
well be that SMTP is the only widespread thing that does this.

This includes a sample config. This works well enough for
openssl s_client -connect -starttls smtp
to work.

TODOs:
- unit tests!
- this probably needs a dedicated config stanza/protobuf, it's just
reusing DownstreamTlsContext now
- what does upstream/client/initiation look like? Is there any way to
get an unmodified smtp client to connect outbound through your service
mesh? At a minimum you'd need to capture the destination IP somehow...
firewall magic?